requirements.txt
googletrans==4.0.0-rc1
streamlit
langdetect
translator  # likely refers to googletrans
PyPDF2
Pillow  # likely refers to PIL (Python Imaging Library)
easyocr
nltk
PyMuPDF
summary
scikit-learn
pyaudio
stt
SpeechRecognition
